Try to fix again!

+++ src/drivers/pci.c
@@ -21,7 +21,7 @@ uint16_t pci_config_read_word(uint8_t bus, uint8_t slot, uint8_t func, uint8_t o
 
     port_dword_out((uint16_t)0xcf8, address);
 
-    uint32_t temp = (uint16_t)((port_dword_in((uint16_t)0xcfc) >> ((offset & 2) * 8)) & 0xffff);
+    uint16_t temp = (uint16_t)((port_dword_in((uint16_t)0xcfc) >> ((offset & 2) * 8)) & 0xffff);
     return temp;
 }
